+++ src/__tests__/useContext.test.tsx
@@ -1,5 +1,5 @@
 import React, { createContext, useContext, FC } from "react";
-import { useTestProxy } from "../useTestProxy";
+import { createTestProxy } from "../createTestProxy";
 
 const TestContext = createContext(0);
 
@@ -8,13 +8,13 @@ const Wrapper: FC<{ val: number }> = ({ val, children }) => (
 );
 
 it("will get the default value", () => {
-  const [prxContext] = useTestProxy(useContext);
+  const [prxContext] = createTestProxy(useContext);
   const res = prxContext(TestContext);
   expect(res).toBe(0);
 });
 
 it("will get the value from the above context", () => {
-  const [prxContext, control] = useTestProxy(useContext, {
+  const [prxContext, control] = createTestProxy(useContext, {
     wrapper: Wrapper,
     props: { val: 2 }
   });