+++ src/index.ts
@@ -1,3 +1,3 @@
-export * from "./useTestProxy";
+export * from "./createTestProxy";
 export { cleanUp } from "./utils";
 export { act } from "react-dom/test-utils";