+++ Changelog.md
@@ -12,3 +12,8 @@
 - Removed useTestHook
 - Added useTestProxy
 - Added act wrap for react
+
+# V2.0.2
+
+- Deprecated useTestProxy to createTestProxy so eslint doesn't pick it up as an error
+- Added fix for resolving with a promise